Updated build gradle to match all versions for spring and HAL etc Updated LovTypeRepostiory for incorrect id data type

+++ src/main/java/com/harman/ebook/vaccination/covid/repository/LovTypeRepostiory.java
@@ -10,6 +10,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  * @created 5/30/2021
  */
 @RepositoryRestResource(collectionResourceRel = "lovtype", path = "lovtype")
-public interface LovTypeRepostiory extends JpaRepository<LovType, Integer> {
+public interface LovTypeRepostiory extends JpaRepository<LovType, Short> {
 
 }
